requirements/base.txt
django-cors-headers==4.7.0
Django==5.2.5
djangorestframework==3.16.1
psycopg2-binary==2.9.10
django-filter==25.1

